Updated item properties and added new items
Significant changes include:
- Updated the properties of existing items in the database, including changing health, mana, and stamina restore percentages from int to double.
- Added detailed descriptions for special elixirs in ItemNames.cs.
- Introduced new types of items: Elixirs that permanently increase player stats and a LevelingItem that increases skill levels.
- Created interfaces for these new item types (IStatIncreaseItem and ILevelingItem).
- Implemented use cases for these new items in the game logic.

+++ DragonsEnd/Items/Status/Interfaces/IRestoreItem.cs
@@ -4,8 +4,8 @@ namespace DragonsEnd.Items.Status.Interfaces
 {
     public interface IRestoreItem : IItem
     {
-        int HealthHealthRestorePercentage { get; set; }
-        int ManaRestorePercentage { get; set; }
-        int StaminaRestorePercentage { get; set; }
+        double HealthHealthRestorePercentage { get; set; }
+        double ManaRestorePercentage { get; set; }
+        double StaminaRestorePercentage { get; set; }
     }
 }
